multi-cloud SnapshotSupportFactory
