+++ service/src/test/java/org/databiosphere/workspacedataservice/annotations/AnnotatedApisTest.java
@@ -3,14 +3,11 @@ package org.databiosphere.workspacedataservice.annotations;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Optional;
-import org.databiosphere.workspacedataservice.activitylog.ActivityLogger;
 import org.databiosphere.workspacedataservice.annotations.DeploymentMode.ControlPlane;
 import org.databiosphere.workspacedataservice.annotations.DeploymentMode.DataPlane;
 import org.databiosphere.workspacedataservice.common.TestBase;
 import org.databiosphere.workspacedataservice.dataimport.protecteddatasupport.ProtectedDataSupport;
 import org.databiosphere.workspacedataservice.dataimport.protecteddatasupport.WsmProtectedDataSupport;
-import org.databiosphere.workspacedataservice.dataimport.snapshotsupport.SnapshotSupportFactory;
-import org.databiosphere.workspacedataservice.dataimport.snapshotsupport.WsmSnapshotSupportFactory;
 import org.databiosphere.workspacedataservice.workspace.DataTableTypeInspector;
 import org.databiosphere.workspacedataservice.workspace.WdsDataTableTypeInspector;
 import org.databiosphere.workspacedataservice.workspacemanager.WorkspaceManagerDao;
@@ -44,13 +41,6 @@ class AnnotatedApisTest extends TestBase {
   // conflicts.
   @TestConfiguration
   static class SpecifyConflictingBeans {
-
-    @Primary
-    @Bean("overrideSnapshotSupportFactory")
-    SnapshotSupportFactory overrideSnapshotSupportFactory(
-        ActivityLogger activityLogger, WorkspaceManagerDao wsmDao) {
-      return new WsmSnapshotSupportFactory(activityLogger, wsmDao);
-    }
 
     @Primary
     @Bean("overrideProtectedDataSupport")
